requirements.txt
azure-core~=1.31.0
Flask~=3.0.3
langchain-core~=0.3.12
langchain-openai~=0.2.2
Werkzeug~=3.0.4